Library Setup
Setting up my code for a library, tweaking and working with various
functions, methods and variables. Pre-defining some commonly used code
so that we may reference it in later works without the need of writing
the same code over and over again.

I have left the JSON.js file as I want to remember it's syntax and
formatting so that I may use it for future projects as an initial
starting point.