Library Changes
Updated library with 9 of the 12 objectives. Fixed some errors with the
string objectives, as well as the key call for arrays. I removed date
difference as I couldn't quite figure it out all the way. Same with
summing total of array, it'd spit out all of the totals each item into
the area, it was weird. Couldn't get that down right but I'm going to
research, tweak and practice more even beyond this class. Will watch a
lot of Lynda.com tutorial series as well.