Array Key Call
Added a way to call an arrays keys but I did see during my research and
testing a way to define a new variable that equals to the array call,
so that you can define new array on the spot and then tab into the
function within the library but haven't got that far yet. As these are
just predefined index's and limited.